LayoutTests/fast/js/reserved-words-expected.txt

This file checks which ECMAScript 3 keywords are treated as reserved words.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


SHOULD BE RESERVED:
PASS isReserved('break') is true
PASS isReserved('case') is true
PASS isReserved('catch') is true
PASS isReserved('class') is true
PASS isReserved('const') is true
PASS isReserved('continue') is true
PASS isReserved('debugger') is true
PASS isReserved('default') is true
PASS isReserved('delete') is true
PASS isReserved('do') is true
PASS isReserved('else') is true
PASS isReserved('enum') is true
PASS isReserved('export') is true
PASS isReserved('extends') is true
PASS isReserved('false') is true
PASS isReserved('finally') is true
PASS isReserved('for') is true
PASS isReserved('function') is true
PASS isReserved('if') is true
PASS isReserved('import') is true
PASS isReserved('in') is true
PASS isReserved('instanceof') is true
PASS isReserved('new') is true
PASS isReserved('null') is true
PASS isReserved('return') is true
PASS isReserved('super') is true
PASS isReserved('switch') is true
PASS isReserved('this') is true
PASS isReserved('throw') is true
PASS isReserved('true') is true
PASS isReserved('try') is true
PASS isReserved('typeof') is true
PASS isReserved('var') is true
PASS isReserved('void') is true
PASS isReserved('while') is true
PASS isReserved('with') is true

SHOULD NOT BE RESERVED:
PASS isReserved('abstract') is false
PASS isReserved('boolean') is false
PASS isReserved('byte') is false
PASS isReserved('char') is false
PASS isReserved('double') is false
PASS isReserved('final') is false
PASS isReserved('float') is false
PASS isReserved('goto') is false
PASS isReserved('implements') is false
PASS isReserved('int') is false
PASS isReserved('interface') is false
PASS isReserved('long') is false
PASS isReserved('native') is false
PASS isReserved('package') is false
PASS isReserved('private') is false
PASS isReserved('protected') is false
PASS isReserved('public') is false
PASS isReserved('short') is false
PASS isReserved('static') is false
PASS isReserved('synchronized') is false
PASS isReserved('throws') is false
PASS isReserved('transient') is false
PASS isReserved('volatile') is false

PASS successfullyParsed is true

TEST COMPLETE
